Add docstrings and remove unused functions
Cropping.py had a function inside that was not being called
by any modules, so I moved it to more_utils.py in case it needed
to be used again in the future. I also added a module docstring
to cropping.py and fixed a few other linting errors.
